LayoutTests/transitions/clip-transition.html

<!DOCTYPE html>

<html>
<head>
  <style type="text/css" media="screen">
    #box {
      position: absolute;
      height: 100px;
      width: 100px;
      background-color: blue;
      clip: rect(10px, 90px, 90px, 10px);
      -webkit-transition: clip 1s linear;
    }

    #box.changed {
      clip: rect(30px, 70px, 70px, 30px);
    }
  </style>
  <script src="transition-test-helpers.js" type="text/javascript" charset="utf-8"></script>
  <script type="text/javascript" charset="utf-8">
    
    const expectedValues = [
      // [time, element-id, property, expected-value, tolerance]
      [0.5, 'box', 'clip', [20, 80, 80, 20], 2],
    ];

    function setupTest()
    {
      document.getElementById('box').className = 'changed';
    }

    runTransitionTest(expectedValues, setupTest, true);
  </script>
</head>
<body>
  <div id="box"></div>
  <div id="result"></div>
</body>
</html>
